♻️ Extract public/driver.go from driver.go
