+++ driver.go
@@ -3,7 +3,6 @@ package sqlslog
 import (
 	"database/sql/driver"
 	"log/slog"
-	"strings"
 )
 
 func wrapDriver(original driver.Driver, logger *SqlLogger) driver.Driver {
@@ -80,28 +79,4 @@ func (w *driverContextWrapper) OpenConnector(dsn string) (driver.Connector, erro
 		lg = lg.With(*attr)
 	}
 	return wrapConnector(origConnector, lg), nil
-}
-
-// DriverOpenErrorHandler returns a function that handles errors from driver.Driver.Open.
-// The function returns a boolean indicating completion and a slice of slog.Attr.
-//
-// # For Postgres:
-// If err is nil, it returns true and a slice of slog.Attr{slog.Bool("success", true)}.
-// If err is io.EOF, it returns true and a slice of slog.Attr{slog.Bool("success", false)}.
-// Otherwise, it returns false and nil.
-func DriverOpenErrorHandler(driverName string) func(err error) (bool, []slog.Attr) {
-	switch driverName {
-	case "postgres":
-		return func(err error) (bool, []slog.Attr) {
-			if err == nil {
-				return true, []slog.Attr{slog.Bool("success", true)}
-			}
-			if strings.ToUpper(err.Error()) == "EOF" {
-				return true, []slog.Attr{slog.Bool("success", false)}
-			}
-			return false, nil
-		}
-	default:
-		return nil
-	}
 }
