+++ driver_test.go
@@ -4,7 +4,6 @@ import (
 	"bytes"
 	"database/sql/driver"
 	"errors"
-	"io"
 	"log/slog"
 	"testing"
 )
@@ -43,34 +42,5 @@ func TestDriverContextWrapperOpenConnector(t *testing.T) {
 		if err == nil {
 			t.Fatal("expected error to be not nil")
 		}
-	})
-}
-
-func TestDriverOpenErrorHandler(t *testing.T) {
-	t.Parallel()
-	t.Run("postgres", func(t *testing.T) {
-		t.Parallel()
-		t.Run("unexpected error", func(t *testing.T) {
-			t.Parallel()
-			eh := DriverOpenErrorHandler("postgres")
-			completed, attrs := eh(errors.New("unexpected error"))
-			if completed {
-				t.Error("expected completed to be false")
-			}
-			if attrs != nil {
-				t.Error("expected attrs to be nil")
-			}
-		})
-		t.Run("io.EOF", func(t *testing.T) {
-			t.Parallel()
-			eh := DriverOpenErrorHandler("postgres")
-			completed, attrs := eh(io.EOF)
-			if !completed {
-				t.Errorf("expected completed to be true")
-			}
-			if attrs == nil {
-				t.Error("expected attrs to be non-nil")
-			}
-		})
 	})
 }
